+++ HomeHQ/Authentication/AuthenticationManager.swift
