+++ HomeHQ/Bill/BillWidgetView.swift
@@ -1,5 +1,5 @@
 //
-//  HouseDutiesWidgetView.swift
+//  BillWidgetView.swift
 //  HomeHQ
 //
 //  Created by Cooper Jacob on 23/8/2023.
@@ -7,16 +7,15 @@
 
 import SwiftUI
 
-struct HouseDutiesWidgetView: View {
+struct BillWidgetView: View {
     
-    @StateObject var viewModel = HouseDutiesViewModel()
     
     var body: some View {
         HStack(alignment: .top) {
-            Text("Today's Tasks")
+            Text("Upcoming Bills")
                 .font(.headline)
         }
-        .frame(height: 80)
+        .frame(height: 150)
         .frame(maxWidth: .infinity)
         .background(Color("ButtonBackground"))
         .cornerRadius(10)
@@ -25,8 +24,8 @@ struct HouseDutiesWidgetView: View {
     }
 }
 
-struct HouseDutiesWidgetView_Previews: PreviewProvider {
+struct BillWidgetView_Previews: PreviewProvider {
     static var previews: some View {
-        HouseDutiesWidgetView()
+        BillWidgetView()
     }
 }
