+++ HomeHQ/ShoppingList/ShoppingListWidgetRowView.swift
