+++ HomeHQ/Utils/ApplicationError.swift
