--- HomeHQ/ViewModel/BillViewModel.swift
@@ -1,13 +1,0 @@
-//
-//  BillViewModel.swift
-//  HomeHQ
-//
-//  Created by Cooper Jacob on 23/8/2023.
-//
-
-import Foundation
-
-// View model for bill view
-class BillViewModel: ObservableObject {
-    
-}
