--- HomeHQ/ViewModel/HouseDutiesViewModel.swift
@@ -1,13 +1,0 @@
-//
-//  HouseDutiesViewModel.swift
-//  HomeHQ
-//
-//  Created by Cooper Jacob on 23/8/2023.
-//
-
-import Foundation
-
-// View model for house duties view
-final class HouseDutiesViewModel: ObservableObject {
-    
-}
